refactor(pdf): isolate equation fragment evidence
